MAAC 3602 - Final Project
This game was designed by myself and a team of 4 other people. We were each given a position in our mock game company and were tasked with completing a game by the end of the semester. My role in this project was that of a producer - to organize deliverables, set deadlines, plan meetings, and do what I could to help our team.